refactor: 공통된 alert 스타일 commonAlert.style로 합쳐주기
