+++ src/components/modal/alert/commonAlert.style.jsx
